+++ src/Shiny.Core/Platforms/Android/PlatformExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 using Shiny.Logging;
 using Android.App;
 using Android.OS;
@@ -12,6 +13,33 @@ namespace Shiny
 {
     public static class PlatformExtensions
     {
+        public static T GetService<T>(this Context context, string name) where T : Java.Lang.Object
+            => (T)context.GetSystemService(name);
+
+        public static bool IsAtLeastAndroid10(this AndroidContext context)
+            => context.IsMinApiLevel(29);
+
+        public static void Execute(this BroadcastReceiver receiver, Func<Task> task)
+        {
+            var pendingResult = receiver.GoAsync();
+            task().ContinueWith(x =>
+            {
+                if (x.IsFaulted)
+                    Log.Write(x.Exception);
+
+                pendingResult.Finish();
+            });
+        }
+
+        static Handler? handler;
+        public static void Dispatch(this Action action)
+        {
+            if (handler == null || handler.Looper != Looper.MainLooper)
+                handler = new Handler(Looper.MainLooper);
+
+            handler.Post(action);
+        }
+
         public static bool IsNull(this Java.Lang.Object obj)
             => obj == null || obj.Handle == IntPtr.Zero;
 
@@ -26,7 +54,6 @@ namespace Shiny
 
         public static void ShinyOnNewIntent(this Activity activity, Intent intent)
             => AndroidShinyHost.TryProcessIntent(intent);
-
 
         public static long ToEpochMillis(this DateTime sendTime)
             => new DateTimeOffset(sendTime).ToUnixTimeMilliseconds();
